requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
torch
scikit-learn
scikit-surprise
requests
python-multipart